Move config parsing to the config_parse submodule
